refactor(ci): extract shared release constants, add missing cache and concurrency
