+++ .github/workflows/tag-staging.yml
@@ -8,6 +8,10 @@ on:
         required: false
         type: string
 
+concurrency:
+  group: tag-staging
+  cancel-in-progress: false
+
 jobs:
   web-e2e:
     name: Web E2E
